requirements.txt
boto3==1.9.34
botocore==1.12.34
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
lxml==4.2.5
Pillow==5.3.0
python-dateutil==2.7.5
python-pptx==0.6.15
requests==2.20.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.24
XlsxWriter==1.1.2
